requirements.txt
aniso8601==8.0.0
asgiref==3.2.7
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.41.0
django-cors-headers==3.3.0
django-countries==6.1.2
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-extensions==0.6.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2-binary==2.8.5
python-decouple==3.3
python-engineio==3.12.1
python-socketio==4.5.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
